+++ app/src/main/java/com/aowen/monolith/feature/search/SearchScreen.kt
@@ -1,7 +1,9 @@
 package com.aowen.monolith.feature.search
 
+import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -15,7 +17,9 @@ import androidx.compose.material.pullrefresh.rememberPullRefreshState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -23,6 +27,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
@@ -35,8 +40,10 @@ import com.aowen.monolith.feature.home.PlayerSearchSection
 import com.aowen.monolith.feature.home.RecentPlayersSection
 import com.aowen.monolith.feature.home.playerdetails.navigation.navigateToPlayerDetails
 import com.aowen.monolith.feature.items.itemdetails.navigation.navigateToItemDetails
+import com.aowen.monolith.feature.matches.navigation.navigateToMatchDetails
 import com.aowen.monolith.ui.components.MonolithAlertDialog
 import com.aowen.monolith.ui.components.MonolithTopAppBar
+import com.aowen.monolith.ui.components.PlayerLoadingCard
 import com.aowen.monolith.ui.components.RefreshableContainer
 import com.aowen.monolith.ui.theme.MonolithTheme
 import com.aowen.monolith.ui.tooling.previews.LightDarkPreview
@@ -62,6 +69,7 @@ internal fun SearchScreenRoute(
         navigateToItemDetails = navController::navigateToItemDetails,
         navigateToHeroDetails = navController::navigateToHeroDetails,
         navigateToBuildDetails = navController::navigateToBuildDetails,
+        navigateToMatchDetails = navController::navigateToMatchDetails,
         navigateBack = navController::popBackStack
     )
 }
@@ -80,6 +88,7 @@ internal fun SearchScreen(
     navigateToPlayerDetails: (String) -> Unit,
     navigateToItemDetails: (String) -> Unit,
     navigateToHeroDetails: (Int, String) -> Unit,
+    navigateToMatchDetails: (String, String) -> Unit,
     navigateToBuildDetails: (Int) -> Unit,
     navigateBack: () -> Unit
 ) {
@@ -155,25 +164,70 @@ internal fun SearchScreen(
                         handleOpenAlertDialog = { clearAllRecentSearchesDialogIsOpen = true }
                     )
                 } else {
-                    HeroSearchSection(
-                        isLoading = uiState.isLoadingItemsAndHeroes,
-                        filteredHeroes = uiState.filteredHeroes,
-                        navigateToHeroDetails = navigateToHeroDetails
-                    )
-                    ItemSearchSection(
-                        isLoading = uiState.isLoadingItemsAndHeroes,
-                        filteredItems = uiState.filteredItems,
-                        navigateToItemDetails = navigateToItemDetails
-                    )
-                    PlayerSearchSection(
-                        uiState = uiState,
-                        handleAddToRecentSearch = handleAddToRecentSearch,
-                        navigateToPlayerDetails = navigateToPlayerDetails
-                    )
-                    BuildsSearchSection(
-                        uiState = uiState,
-                        navigateToBuildDetails = navigateToBuildDetails,
-                    )
+                    if (uiState.foundMatch != null) {
+                        MatchSearchSection(
+                            isLoading = uiState.isLoadingMatchSearch,
+                            foundMatch = uiState.foundMatch,
+                            navigateToMatchDetails = navigateToMatchDetails
+                        )
+                    } else {
+                        Column(
+                            verticalArrangement = Arrangement.spacedBy(8.dp)
+                        ) {
+                            Row(
+                                modifier = Modifier.fillMaxWidth(),
+                                horizontalArrangement = Arrangement.SpaceBetween,
+                                verticalAlignment = Alignment.CenterVertically
+                            ) {
+                                Text(
+                                    text = "Match",
+                                    color = MaterialTheme.colorScheme.secondary,
+                                    style = MaterialTheme.typography.titleMedium
+                                )
+                            }
+                            AnimatedContent(
+                                targetState = uiState.isLoadingMatchSearch,
+                                label = ""
+                            ) { isLoadingSearch ->
+                                if (isLoadingSearch) {
+                                    Column(
+                                        modifier = Modifier.fillMaxWidth(),
+                                        horizontalAlignment = Alignment.CenterHorizontally,
+                                        verticalArrangement = Arrangement.spacedBy(8.dp)
+                                    ) {
+                                        PlayerLoadingCard(
+                                            titleWidth = 100.dp,
+                                            subtitleWidth = 75.dp
+                                        )
+                                    }
+                                } else {
+                                    Text(
+                                        text = "No match found.",
+                                        color = MaterialTheme.colorScheme.secondary
+                                    )
+                                }
+                            }
+                        }
+                        HeroSearchSection(
+                            isLoading = uiState.isLoadingItemsAndHeroes,
+                            filteredHeroes = uiState.filteredHeroes,
+                            navigateToHeroDetails = navigateToHeroDetails
+                        )
+                        ItemSearchSection(
+                            isLoading = uiState.isLoadingItemsAndHeroes,
+                            filteredItems = uiState.filteredItems,
+                            navigateToItemDetails = navigateToItemDetails
+                        )
+                        PlayerSearchSection(
+                            uiState = uiState,
+                            handleAddToRecentSearch = handleAddToRecentSearch,
+                            navigateToPlayerDetails = navigateToPlayerDetails
+                        )
+                        BuildsSearchSection(
+                            uiState = uiState,
+                            navigateToBuildDetails = navigateToBuildDetails,
+                        )
+                    }
                 }
             }
         }
@@ -199,6 +253,7 @@ private fun SearchScreenPreview() {
             navigateToItemDetails = {},
             navigateToHeroDetails = { _, _ -> },
             navigateToBuildDetails = {},
+            navigateToMatchDetails = { _, _ -> },
             navigateBack = {}
         )
     }