+++ app/src/main/java/com/aowen/monolith/network/UserPreferencesManager.kt
@@ -12,8 +12,10 @@ val Context.userPreferencesDataStore by preferencesDataStore(name = "user_prefer
 
 interface UserPreferencesManager {
     val console: Flow<Console>
+    val claimedPlayerName: Flow<String?>
     val hasSkippedOnboarding: Flow<Boolean>
     suspend fun saveConsole(console: Console)
+    suspend fun saveClaimedPlayerName(playerName: String?)
     suspend fun setSkippedOnboarding()
 }
 
@@ -21,13 +23,18 @@ class UserPreferencesManagerImpl(private val context: Context) : UserPreferences
 
     companion object {
         private val CONSOLE_KEY = stringPreferencesKey("console")
+        private val CLAIMED_PLAYER_NAME_KEY = stringPreferencesKey("claimed_player_name")
         private val SKIPPED_ONBOARDING_KEY = stringPreferencesKey("skipped_onboarding")
-        private val SYNCED_SAVED_CONTENT_KEY = stringPreferencesKey("synced_saved_content")
     }
 
     override val console: Flow<Console> = context.userPreferencesDataStore.data
         .map { prefs ->
             prefs[CONSOLE_KEY]?.let { Console.valueOf(it) } ?: Console.PC
+        }
+
+    override val claimedPlayerName: Flow<String?> = context.userPreferencesDataStore.data
+        .map { prefs ->
+            prefs[CLAIMED_PLAYER_NAME_KEY]
         }
 
     override val hasSkippedOnboarding: Flow<Boolean> = context.userPreferencesDataStore.data
@@ -38,6 +45,18 @@ class UserPreferencesManagerImpl(private val context: Context) : UserPreferences
     override suspend fun saveConsole(console: Console) {
         context.userPreferencesDataStore.edit { prefs ->
             prefs[CONSOLE_KEY] = console.name
+        }
+    }
+
+    override suspend fun saveClaimedPlayerName(playerName: String?) {
+        if(playerName == null) {
+            context.userPreferencesDataStore.edit { prefs ->
+                prefs.remove(CLAIMED_PLAYER_NAME_KEY)
+            }
+        } else {
+            context.userPreferencesDataStore.edit { prefs ->
+                prefs[CLAIMED_PLAYER_NAME_KEY] = playerName
+            }
         }
     }